pipeline/requirements.txt
pytest
pylint
boto3
pandas
python-dotenv
redshift_connector